backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
gunicorn>=20.0
PyJWT>=2.0
Werkzeug>=2.0
